refactor: tanstack query의 queryKey constant로 작성
