+++ src/constants/queryKey.ts
@@ -1,0 +1,2 @@
+export const AUTH_KEY = 'AUTH';
+export const DIARY_KEY = 'DIARY';
