+++ src/features/auth/apis/queries/useGetAuthCode.ts
@@ -1,5 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
-import { getAuthCode, AUTH_KEY } from '@/features/auth/apis/index.ts';
+import { AUTH_KEY } from '@/constants/queryKey.ts';
+import { getAuthCode } from '@/features/auth/apis/index.ts';
 import { GoogleLoginUrl } from '@/features/auth/apis/interfaces.ts';
 
 const useGetAuthCode = () =>
