+++ src/features/diaryList/apis/mutations/useDeleteFile.ts
@@ -1,7 +1,8 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { DIARY_KEY, deleteFile } from '@/features/diaryList/apis/index.ts';
+import { deleteFile } from '@/features/diaryList/apis/index.ts';
 import { DriveFile, File } from '@/features/diaryList/apis/interfaces.ts';
 import { AxiosError } from 'axios';
+import { DIARY_KEY } from '@/constants/queryKey.ts';
 
 const useDeleteFile = () => {
   const queryClient = useQueryClient();
