+++ src/features/diaryList/apis/mutations/usePatchFile.ts
@@ -1,6 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
-import { DIARY_KEY, patchFile } from '@/features/diaryList/apis/index.ts';
+import { patchFile } from '@/features/diaryList/apis/index.ts';
 import { PatchFileArgs } from '@/features/diaryList/apis/interfaces.ts';
+import { DIARY_KEY } from '@/constants/queryKey.ts';
 
 const usePatchFile = () =>
   useMutation({
