+++ src/features/diaryList/apis/queries/useGetFile.ts
@@ -1,4 +1,5 @@
-import { DIARY_KEY, getFile } from '@/features/diaryList/apis/index.ts';
+import { DIARY_KEY } from '@/constants/queryKey.ts';
+import { getFile } from '@/features/diaryList/apis/index.ts';
 import { useQuery } from '@tanstack/react-query';
 
 const useGetFile = (id: string, accessToken: string | undefined) =>
