+++ src/features/diaryList/apis/queries/useGetFileList.ts
@@ -1,6 +1,7 @@
-import { DIARY_KEY, getFileList } from '@/features/diaryList/apis/index.ts';
+import { getFileList } from '@/features/diaryList/apis/index.ts';
 import { useQuery } from '@tanstack/react-query';
 import { DriveFile } from '@/features/diaryList/apis/interfaces.ts';
+import { DIARY_KEY } from '@/constants/queryKey.ts';
 
 const useGetFileList = (accessToken: string | undefined) =>
   useQuery<DriveFile>({
