+++ src/features/diaryList/apis/queries/useGetFileMetaData.ts
@@ -1,4 +1,5 @@
-import { DIARY_KEY, getFileMetaData } from '@/features/diaryList/apis/index.ts';
+import { DIARY_KEY } from '@/constants/queryKey.ts';
+import { getFileMetaData } from '@/features/diaryList/apis/index.ts';
 import { MetaData } from '@/features/diaryList/apis/interfaces.ts';
 import { useQuery } from '@tanstack/react-query';
